DO NOT MERGE: WM: Only allow system to use NO_INPUT_CHANNEL.
NO_INPUT_CHANNEL is a hidden WM flag that allows creation of a window
without an input channel. Unfortunately in releases prior to Android R
this would allow creation of a Window which will not be known to the
InputDispatcher at all. This means that the logic generating
FLAG_OBSCURED will work and a window will be able to overlay another
window without the overlayed window being notified. In Android R and
later this isn't a problem as the InputDispatcher is informed of all
windows, input channel or not. For past Android releases, this patch
disables NO_INPUT_CHANNEL for use outside of the WM.

Bug: 152064592
Test: Existing tests pass
(cherry picked from commit 9661bf7a40d33470bcd35f15758ab71c7c635c63)

[basilgello: Back-port to 14.1:
 - SYSTEM_UID -> Process.SYSTEM_UID]

Signed-off-by: Vasyl Gello <vasek.gello@gmail.com>
Change-Id: I7e1f45cba139eab92e7df88d1e052baba0ae2cc6